+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2019-07-10
+### Added
+- Support for conversion to String representations including hexadecimal, base 64 and encoded using `StringRepresentable`.
+- Initializers for initializing `Data` or `String` from a hexadecimal String.
+
 ## [1.1.3] - 2019-07-08
 ### Added
 - Support for AES-192 and AES-256.